+++ CHANGELOG.md
@@ -8,11 +8,13 @@ GAMSPy 0.12.1
   - __pow__ returns sqrt instead of rPower if the exponent is 0.5.
   - Deprecate delayed_execution mode.
   - Replace pylint, flake8 and black with ruff.
+  - Implement /api/auth -> post, /api/auth/login -> post and /api/auth/logout -> post for GAMS Engine.
 - Testing
   - Replace cta PandasExcelReader and PandasExcelWriter with new ExcelReader and ExcelWriter from GAMS Connect correspondingly. 
   - Add a new model (Nurses) to the model library and the Notebook examples.
   - Add an AC optimal power flow (ACOPF) model to the model library.
   - Add a test to verify the generated string for power where the exponent is 0.5.
+  - Add tests for /api/auth.
 - Documentation
   - Remove FAQ about Google Colab (it is resolved) and add FAQ about Windows Defender.
   - Remove documentation for delayed execution mode.